# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TravelBook API — v1 vertical slice
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /planner/build:
    post:
      operationId: buildPlan
      tags: [planner]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildPlanInput" }
      responses:
        "202":
          description: Build started — passId + buildId for the SSE stream
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildPlanAccepted" }
  /planner/{passId}/build-stream:
    parameters:
      - name: passId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: streamPlanBuild
      tags: [planner]
      description: |
        Server-Sent Events. Each event is `{ kind, state, item }` per the
        progressing build. Stream closes when the build completes or errors.
        The generated client will return a Response; consumers should subscribe
        with the SSE-aware fetch wrapper.
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
  /planner/{passId}:
    parameters:
      - name: passId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getPlan
      tags: [planner]
      responses:
        "200":
          description: Pass + items + member counts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassWithItems" }
  /planner/{passId}/items/{itemId}:
    parameters:
      - name: passId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: itemId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: patchItem
      tags: [planner]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchItemInput" }
      responses:
        "200":
          description: Patched item + reshuffled downstream
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassWithItems" }
  /planner/{passId}/items/{itemId}/swap:
    parameters:
      - name: passId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: itemId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: swapItem
      tags: [planner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alternateId]
              properties:
                alternateId: { type: string, format: uuid }
      responses:
        "200":
          description: Swapped + reshuffled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassWithItems" }
  /planner/{passId}/book-all:
    parameters:
      - name: passId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: bookAll
      tags: [planner]
      responses:
        "200":
          description: Pass minted (state → upcoming, items confirmed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassWithItems" }
  /items/{itemId}/alternates:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listAlternates
      tags: [planner]
      responses:
        "200":
          description: Alternate options for this slot
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Alternate" }

  # ───────────────────────────────────────── Passes
...
components:
  schemas:
...
    PassState:
      type: string
      enum: [draft, live, upcoming, archived]
    PassVisibility:
      type: string
      enum: [private, link, friends, public]
    ItineraryKind:
      type: string
      enum: [flight, stay, activity, transit, dining, insurance, visa, event, entertainment]
    ItineraryState:
      type: string
      enum: [draft, confirmed, past, shifted, swapped, cancelled]
    Pass:
      type: object
      required: [id, userId, passNo, code, title, startsOn, endsOn, state, totalCost, currency, visibility, openSeats, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        passNo: { type: integer }
        code: { type: string }
        title: { type: string }
        coverImageUrl: { type: string, nullable: true }
        startsOn: { type: string }
        endsOn: { type: string }
        state: { $ref: "#/components/schemas/PassState" }
        totalCost: { type: string }
        currency: { type: string }
        aiPrompt: { type: string, nullable: true }
        aiBuiltAt: { type: string, nullable: true }
        visibility: { $ref: "#/components/schemas/PassVisibility" }
        openSeats: { type: integer }
        seatsClosesAt: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
    ItineraryItem:
      type: object
      required: [id, passId, sortAt, kind, state, title, subtitle, currency, data]
      properties:
        id: { type: string, format: uuid }
        passId: { type: string, format: uuid }
        sortAt: { type: string }
        kind: { $ref: "#/components/schemas/ItineraryKind" }
        state: { $ref: "#/components/schemas/ItineraryState" }
        providerId: { type: string, format: uuid, nullable: true }
        externalRef: { type: string, nullable: true }
        title: { type: string }
        subtitle: { type: string }
        code: { type: string, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        cost: { type: string, nullable: true }
        currency: { type: string }
        alternatesCount: { type: integer }
        data:
          type: object
          additionalProperties: true
    Alternate:
      type: object
      required: [id, itemId, rank, data, tags]
      properties:
        id: { type: string, format: uuid }
        itemId: { type: string, format: uuid }
        rank: { type: integer }
        data:
          type: object
          additionalProperties: true
        tags:
          type: array
          items: { type: string }
        deltaCost: { type: string, nullable: true }
        deltaMinutes: { type: integer, nullable: true }
        score: { type: number, nullable: true }
    PassMember:
      type: object
      required: [userId, role, joinedAt]
      properties:
        userId: { type: string, format: uuid }
        handle: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        role:
          type: string
          enum: [owner, holder, buddy, requested]
        joinedAt: { type: string }
    PassWithItems:
      type: object
      required: [pass, items, members]
      properties:
        pass: { $ref: "#/components/schemas/Pass" }
        items:
          type: array
          items: { $ref: "#/components/schemas/ItineraryItem" }
        members:
          type: array
          items: { $ref: "#/components/schemas/PassMember" }
        fitCheck:
          $ref: "#/components/schemas/FitCheck"
          nullable: true
...
    BuildPlanInput:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
        dates:
          type: object
          nullable: true
          properties:
            start: { type: string }
            end: { type: string }
        budgetUsd: { type: number, nullable: true }
        travelers: { type: integer, default: 1 }
        useDocuments:
          type: array
          items: { type: string, format: uuid }
    BuildPlanAccepted:
      type: object
      required: [passId, buildId]
      properties:
        passId: { type: string, format: uuid }
        buildId: { type: string }
    PatchItemInput:
      type: object
      properties:
        state:
          type: string
          enum: [confirmed, cancelled]
        sortAt: { type: string }
        data:
          type: object
          additionalProperties: true
...
    FitCheck:
      type: object
      required: [score, breakdown]
      properties:
        score: { type: number }
        breakdown:
          type: object
          properties:
            pace:
              type: string
              enum: [match, soft, mismatch]
            budget:
              type: string
              enum: [match, soft, mismatch]
            mornings:
              type: string
              enum: [match, soft, mismatch]
            diet:
              type: string
              enum: [match, soft, mismatch]
        costPreview:
          type: array
          items:
            type: object
            properties:
              kind: { $ref: "#/components/schemas/ItineraryKind" }
              cost: { type: string }
...